WORKING SEARCH IMPLEMENTATION (scales poorly)
Working server-side recursive search implementation.  It is still a
direct operation w/o a caching implementation, so it doesn't scale very
well and can timeout when searching in a very large structure (same kind
of limitations as would be expected running a 'find' command on a large
dir). This is mostly a reference implementation and proof-of-concept in
its current state. Later versions will likely involve Filesystem driver
APIs (this implementation is up at the module/view layer) which utilize
some kind of caching mechanism, etc...